requirements.txt
alembic==1.12.0
fastapi==0.103.2
passlib==1.7.4
pydantic==2.4.2
python-jose==3.3.0
python-multipart==0.0.6
psycopg2==2.9.9
SQLAlchemy==2.0.22
uvicorn==0.23.2